feat: move `ChainIndexer` and `ChainStream` traits to `mpc-indexer-core`
