+++ chain-signatures/indexer-core/Cargo.toml
@@ -4,5 +4,10 @@ edition = "2024"
 version.workspace = true
 
 [dependencies]
+# TODO: replace anyhow with thiserror
+anyhow.workspace = true
 async-trait.workspace = true
+futures-util.workspace = true
 mpc-primitives.workspace = true
+tracing.workspace = true
+tokio.workspace = true
