--- chain-signatures/indexer-core/src/stream.rs
@@ -1,1 +1,0 @@
-
