+++ chain-signatures/node/src/indexer_canton/stream.rs
@@ -1,12 +1,11 @@
 use crate::protocol::Chain;
 use crate::rpc::CantonClient;
-use crate::stream::{ChainIndexer, ChainStream};
 
 use alloy::primitives::keccak256;
 use async_trait::async_trait;
 use futures_util::stream::{self, SplitSink, SplitStream};
 use futures_util::{SinkExt, StreamExt};
-use mpc_indexer_core::{ChainTelemetry, StateManager};
+use mpc_indexer_core::{ChainTelemetry, StateManager, ChainIndexer, ChainStream};
 use mpc_primitives::{
     ChainEvent, RespondBidirectionalEvent, ScalarExt, Signature, SignatureRespondedEvent,
 };
