+++ chain-signatures/node/src/stream/mod.rs
@@ -4,7 +4,7 @@ pub mod pipeline;
 use crate::backlog::Backlog;
 use crate::mesh::MeshState;
 use crate::node_client::NodeClient;
-use crate::protocol::{Chain, Sign};
+use crate::protocol::Sign;
 use crate::rpc::{ContractStateWatcher, RpcChannel};
 use crate::stream::ops::{
     process_block_event, process_execution_confirmed, process_respond_bidirectional_event,
@@ -13,11 +13,8 @@ use crate::stream::ops::{
 };
 pub use crate::stream::pipeline::ChainPipeline;
 
-use async_trait::async_trait;
-use futures_util::Stream;
-use mpc_indexer_core::ChainTelemetry;
+use mpc_indexer_core::{ChainTelemetry, ChainStream, ChainIndexer};
 use mpc_primitives::{ChainEvent, CheckpointDigest};
-use std::time::Duration;
 use tokio::sync::{mpsc, watch};
 
 pub const CHAIN_EVENT_STREAM_SIZE: usize = 16384;
@@ -31,60 +28,6 @@ pub enum ChainStreaming {
 
 pub fn channel() -> (mpsc::Sender<ChainEvent>, mpsc::Receiver<ChainEvent>) {
     mpsc::channel(CHAIN_EVENT_STREAM_SIZE)
-}
-
-#[async_trait]
-pub trait ChainIndexer: Send + 'static {
-    const CHAIN: Chain;
-    type Block: Send;
-    type Iter: Stream<Item = Self::Block> + Send + Unpin + 'static;
-
-    const RETRY_DELAY: Duration = Duration::from_millis(500);
-
-    async fn livestream(&mut self) -> anyhow::Result<Option<u64>> {
-        Ok(None)
-    }
-
-    async fn notify_catchup_completed(&mut self) -> anyhow::Result<()> {
-        Ok(())
-    }
-
-    async fn catchup_range(&self, anchor_height: u64) -> Self::Iter;
-
-    async fn process_catchup(&mut self, item: &Self::Block) -> anyhow::Result<()> {
-        let _ = item;
-        Ok(())
-    }
-
-    async fn next(&mut self) -> Option<Self::Block> {
-        None
-    }
-
-    async fn process(&mut self, block: &Self::Block) -> anyhow::Result<()> {
-        let _ = block;
-        Ok(())
-    }
-
-    /// Process the next block, return true for success, false for shutdown.
-    async fn process_next_block(&mut self) -> bool {
-        let Some(block) = self.next().await else {
-            return false;
-        };
-
-        while let Err(err) = self.process(&block).await {
-            tracing::warn!(?err, "live block processing failed; retrying");
-            tokio::time::sleep(Self::RETRY_DELAY).await;
-        }
-        true
-    }
-}
-
-#[async_trait]
-pub trait ChainStream: Send + 'static {
-    type Indexer: ChainIndexer + Send;
-
-    async fn start(&mut self) -> anyhow::Result<Self::Indexer>;
-    async fn next_event(&mut self) -> Option<ChainEvent>;
 }
 
 /// Shared indexer loop: recovers backlog then processes events from the stream
@@ -222,10 +165,12 @@ mod tests {
     use crate::rpc::{ContractStateWatcher, RpcAction, RpcChannel};
     use crate::storage::checkpoint_storage::CheckpointStorage;
     use crate::util::current_unix_timestamp;
+    use async_trait::async_trait;
     use k256::{AffinePoint, Scalar};
     use mockito::Server;
-    use mpc_indexer_core::{NoopChainTelemetry, StateManager};
+    use mpc_indexer_core::{NoopChainTelemetry, StateManager, ChainIndexer, ChainStream};
     use mpc_primitives::{
+        Chain,
         CheckpointDigest, IndexedSignRequest, SignArgs, SignId, Signature, SignatureRespondedEvent,
     };
     use near_primitives::types::AccountId;
@@ -300,7 +245,7 @@ mod tests {
                 }
             }
 
-            #[async_trait]
+            #[async_trait::async_trait]
             impl ChainStream for $stream {
                 type Indexer = $indexer;
