+++ chain-signatures/node/src/stream/pipeline.rs
@@ -2,8 +2,9 @@ use crate::backlog::Backlog;
 use crate::mesh::MeshState;
 use crate::node_client::NodeClient;
 use crate::protocol::Chain;
-use crate::stream::{ChainIndexer, ChainStreaming};
+use super::ChainStreaming;
 use futures_util::StreamExt;
+use mpc_indexer_core::ChainIndexer;
 use mpc_primitives::CheckpointDigest;
 use near_account_id::AccountId;
 use tokio::sync::watch;
